+++ projects/start-tunnel/web/src/app/services/api/live-api.service.ts
@@ -225,7 +225,7 @@ export class LiveApiService extends ApiService {
             ...options,
             headers: {
               ...options.headers,
-              ...this.authKeys.signRpcHeaders(options),
+              ...(await this.authKeys.signRpcHeaders(options)),
             },
           },
       urlOverride,